refactor lazy_ir helpers into api/lazy.py
